Fix calculation of CPU reservations
Previously, the CPU reservations in Grafana were reported incorrectly.
For some reason, these were not correctly counted when using an InfluxDB
query, but were correctly calculated with a very similar query in
Prometheus. Create a pre-calculated query in Prometheus for CPU
reservations and use that in the default InfluxDB dashboard instead for
showing reserved CPU capacity.