Compute node stack flavor update policy
Allow setting the flavor update policy on compute node stacks. This can
be set to either RESIZE or REPLACE according to Heat documentation. The
default is kept as RESIZE to be on the safe side. If the policy is not
specifically given by the user then everything should remain as before.

Some flavor types cannot be resized, in which case it can be useful to
be able to set REPLACE as the policy instead. This should obviously be
used with care, as changing the flavor for all nodes in a resource group
would cause replacement of all compute nodes at the same time. A
possible use case would be to create a second compute node stack
alongside the existing one to have a place to evacuate pods before
replacing the old compute nodes with bigger ones all at once.