keepalived enhancements
This commit adds an option to configure the virtual router id used in
VRRP between the load balancers. This is useful when there are multiple
VRRP instances in the network.

We also change the variable 'keepalived_vip_mask' to
'keepalived_vip_broadcast' to match the purpose. There is a fallback
mechanism for using the old variable if the inventory has not been
converted yet.