deployment with public IP network
This commit adds support for deploying POC on OpenStack systems that do
not have support for floating IPs, but use a network with public IPs
available directly.

Changes:

- introduce separate Heat stacks to be able to provision VMs with two
  interfaces
- refactor the heat stack file directory layout
- add support for configuring eth1
- add an iptables rule to allow keepalived traffic on lbs on eth1
- increased the ssh timeout in ansible.cfg, the initial connection
  to nodes with public IPs on is timing out otherwise

Notes:
An alternative approach was tried first with allocating ports from the
public network in ansible (instead of Heat) and attaching those to VMs.
This  turned out to be slow and cumbersome. The VMs needed a cold power
cycle to get security groups to work correctly.