enhance rolling_host_update.bash
This commit fixes an issue with scaleup playbook name by introducing
an optional parameter for defining the version for scaleup playbook
and defaults to '3.9', which is currently used.

It also removes the option for specifying the image for rebuilds.
Changing the image is not usually needed and might be dangerous if the
new image differs from the specified in the Heat stack.

It also runs package updates as part of rebuild to get up-to-date host
to run openshift-ansible on.