InfluxDB scalability enhancements
Add support for whitelisting Prometheus metrics for long term storage in
InfluxDB. This whitelist is in Prometheus' remote_write config.
InfluxDB's downsampling config need not be changed.

To monitor how the InfluxDB database grows, add a new Grafana dashboard
for InfluxDB metamonitoring. Crucially, this new dashboard shows the
number of series in InfluxDB, which is an important indicator for
increasing resource needs. At the same time, reduce repetition in the
deploy_monitoring playbook and add a new datasource (InfluxDB _internal
DB) that is needed by the InfluxDB metamonitoring dashboard.

Bump up the max-series-per-database setting in InfluxDB from 1000000 to
2000000. While it is unlikely that this limit will be reached soon with
the whitelisting config, the hardware we have should be able to handle
more than 1000000 series in InfluxDB.

Finally, remove some environment specific config (hostnames, names of
downsampled InfluxDB databases) from the Grafana boards that are
automatically installed.