Set minimum to zero for Grafana dashboard graphs
Some of the capacity planning graphs we have had their minimum Y axis
value set to "auto" which would in some cases make it look as if we had
more free space than we actually have. Explicitly set the minimum for
the Y axis to zero so that the graphs are better at visualizing actual
capacity used and left.