default ldap_bind_account_password_vault
Previously we had the ldap bind account password commonly defined in
all environments. After refactoring it to be environment specific, it
became apparent that even if ldap auth is not deployed, the variable was
needed in the task configuring the auth plugins here:

playbooks/roles/openshift_auth_providers/tasks/main.yml

Here we introduce a fallback default variable for
ldap_bind_account_password_vault with a string "UNDEFINED" as the value,
so that if it ever ended up in a config it would be easy to detect
during troubleshooting.