CI/CD pipeline without deployer image build
Modify the GitLab CI/CD pipeline so that we no longer build the Docker
image during the pipeline. Instead of building the image itself, the
pipeline now takes a pre-built image and uses that directly. This is
done to reduce flakes caused by things like Alpine mirror issues that
have been quite common. It should also speed up the pipeline in the
normal scenario as well, as there are fewer steps. It also greatly
simplifies .gitlab-ci.yml and makes it easier to read.

Update README.md with instructions on the new way to use the
poc-deployer image in the pipeline. At the same time, reorganize the
document to have a more logical ordering and fix some markdown style
issues.

Refactor init_env.bash slightly so that repo cloning is done in a
function. This is a leftover from trying to clone other repositories
during init_env.bash that is now only used for openshift-ansible, but I
decided to leave it in the script anyway since it works as a generic
improvement as well. In the future, the function can easily be used to
clone more than just openshift-ansible as well in case stand alone
deployment containers (== no mounts) are implemented.