Make it possible to toggle OS version locking
Previously when creating additional OS repo configs, state would always
be set to "present" and the configs could not be removed. Add the state
option and make it so that it is toggled based on the
lock_os_minor_version setting.

Having the same repository defined twice - even if one definition sets
enabled=0 causes problems for package installations.