More lenient Rocket.Chat template liveness probe
The initial delay for the liveness probe for the Rocket.Chat application
in the Rocket.Chat template was set to 30 seconds. It looks like this is
not quite enough time for the container to start serving initially,
which caused it to go into a crash loop most of the time. To give the RC
container more time to start, give it 60 seconds to start up instead of
30.