Fix ordering of namespace quota rules
According to the OpenShift documentation, the namespace rules under
ProjectRequestLimitConfig are evaluated in order, and the first one to
match a user is used. This means that if the default rule is at the top,
this matches *all* users and is applied to everyone regardless of any
labels set.

Move the default rule for the number of namespaces to the bottom so that
users with the right label set are evaluated using a different rule that
is higher up in the list.