Common compute node anti-affinity server group
Create a new common server group for compute nodes from multiple Heat
stacks. The UUID of this new stack is added to the outputs from the base
stack and is optionally used when creating compute node stacks. Whether
to use the common server group for compute nodes in a Heat stack can be
selected using a new boolean parameter called use_common_server_group.
By default this parameter is false in which case the Heat stack behaves
exactly the same as before and creates a dedicated server group for
nodes created in the compute node stack. This makes the change backwards
compatible so that existing stacks are not left in a precarious state.

The new server group is named based on the name of the environment. Add
a comment in the Heat template that warns against renaming this server
group after it has been created and taken into use.

Document usage of the new common server group feature in
docs/compute_node_stacks.md and also add generic documentation on how
compute node stacks can be managed with poc. Update the main README.md
file with a table of contents so that information is easier to find
there.