Configurable project default quota
The values for the default quota in a project were hardcoded in the
project-request.yaml template. Make them configurable using variables of
the form default_project_*. These can be set on a per-environment basis
depending on what is needed in each environment. We can also update
existing environments with a different default quota if this is seen as
necessary.

Default are set for all values to be the same as the previous hard coded
values to avoid having to set these variables for all existing
environments.