Fix node scaledown for OKD 3.10
Starting from OKD 3.10, regular compute nodes always have pods from a
DaemonSet present on them. Draining these nodes will thus always fail
unless you set the --ignore-daemonsets parameter for the drain command.
Do this in the scaledown playbook.

Also update our documentation to clarify that you can only remove nodes
from the end of the Heat stack. This clarification is now before the
instructions on listing nodes in nodes_to_remove.yaml so that the reader
is aware of this limitation when listing the nodes to remove.