Fix SSH config generation to work with more nodes
We previously generated a very long line of hostnames and IP addresses
for common SSH config for cluster hosts. When enough nodes were added to
the cluster, SSH failed to parse this long line correctly and would
split it, which would cause a parsing error.

Since we run everything from a deployment container these days, we no
longer need to be as explicit in the SSH config generation. Using
wildcards is fine. Instead of a long list of hosts, just use two
wildcard entries:

 - {{ cluster_name }}-*
 - {{ openshift_network_prefix }}.*