Flowdock notifications on playbook start/end
Write an Ansible callback plugin that notifies Flowdock when a playbook
finishes if a Flowdock API token is found. If no token is found then the
callback plugin does nothing.

Modify .gitlab-ci.yml so that the new callback plugin gets the data it
needs during pipeline runs. A new mechanism for passing data to Ansible
from the GitLab pipeline is established for this purpose. As there were
changes to the deployer image, also update the image tag for that.

Update the requirements for the poc-deployer Docker image and add:

 - bpython and bpdb for debugging callback plugins and other Python code
 - python-flowdock for sending notifications to Flowdock

Modify the run_playbook.bash script so that it only prints the first
argument it gets to the console. If we would pass the API token via an
argument to this script, we would not want this to be printed into our
pipeline logs.

Add the new callback plugin in ansible.cfg so that it gets called.

Finally, a separate flowdock task is added in site no send notices of
when a playbook run starts. This turned out to be difficult to do in the
callback plugin as it does not have the necessary context to send a
useful notification at the start of a playbook run.

The reason for not using the flowdock module for everything and instead
writing a dedicated callback plugin was that this was the only way to
reliably get notifications also on failed playbook runs.

Document the implementation under docs/flowdock_callback_plugin.md.