Document how and when to use vault pass script
In the main README.md file under "Provisioning", give instructions on
when and how to run the read_vault_pass_from_clipboard.bash script. Move
the chapter about provisioning right after the prerequisites chapter to
make the docs easier to follow when you are trying to use poc for the
first time.