Allow setting alt names for hosts in /etc/hosts
Sometimes we might want to have an alternate name for a host. This is
the case with e.g. GlusterFS nodes that have already started using an
old name previously and where renaming could cause disruptions to
service.

Make it possible to set this alternate name in /etc/hosts without
affecting the actual hostname of the host. Also make the name with the
poc_override_domain_suffix added optional.

Set the default for poc_override_domain_suffix as an empty string so
that the default behaviour of the cluster_common role is to not add
additional entries in /etc/hosts for an additional suffix.

Due to the way variable evaluation works in Ansible, this change means
that we need to move the cluster_common role on the bastion until after
we've gathered facts for all other hosts. Otherwise we won't have access
to ansible_* facts for setting the alternate names of hosts in the
bastion's /etc/hosts file.