Add plain app domain route and redirect for it
In setup_default_www_app.yml, change the items in route and redirect creation tasks to objects to support per-item decision on whether the item should be applied in the task. Var openshit_app_domain does not exist for every env, so it should not be added to as route to those envs.