Add support for setting up the Cinder CSI plugin
Add a new role called csi_cinder that can be used to install the
necessary components for running the Cinder CSI plugin. This role is
called in the extra_oso_config playbook that's previously been used to
e.g. set up storage classes.

Some defaults are declared for the role such as the namespace in which
the components are to be deployed (the default is a new dedicated
namespace called csi-cinder) and a set of images to use.

This addition is mainly based on the instructions in OKD
documentation here:

https://docs.okd.io/latest/install_config/persistent_storage/persistent_storage_csi.html

Automatic tests are added that check whether deployments work with both
the old GlusterFS storage class and the new Cinder based storage class.
There is also a separate test that does not specify a storage class and
thus will ensure that there is a working default storage class. The
deployment check Nagios script is amended to make these tests possible.