Dedicated IPs for namespaces
* Add "compute_allowed_address_pairs_cidrs" to the compute nodes
Heat templated in order to allow the compute node interfaces to
send traffic using multiple IP addresses. This is an array which
can support multiple CIDRs.

* Define the egress-ips Heat template which includes the port/
floatingIpAssociation pair, in order to map private cluster IPs to
public routable floating IPs. This was not necessary for clusters
with public subnets as those can directly have public routable IPs.
The template is Jinja2 based, as looping over multiple resources
in Heat is not supported. It goes through the `egress_ips` list
and for each entry; it creates a port with fixed a IP and
associates it to a floating IP.

* Introduce a new role in poc: `dedicated_namespace_ip` which
takes care of making sure that the project which has a dedicated
IP exists, and the namespace owner matches the admin user from the
inventory. It also allocates the correct private IP (which
corresponds to a public IP) to the namespace. Moreover, it allows
all of the OpenShift compute nodes to host customer IPs
from the allocated CIDRs. So that if a compute node goes down, a
different one will pickup the IP address.

* Delete the egress-ip stack in deprovision if it exists.

* Add the namespace egress ip test to bats tests.

* Setup the egress IPs ports and namespaces as part of Gitlab CI
for oso-ci and oso-qa for running tests which check the egress IP
and make sure that it matches the floating IP from the inventory.