k8s module fix by changing kubeconfig CA data
The k8s Ansible module
(https://docs.ansible.com/ansible/latest/modules/k8s_module.html) does
not work on multimaster installations with the
certificate-authority-data value that is configured by default. It will
throw an SSL verification error in these cases.

Change certificate-authority-data in root's and cloud-user's kubeconfig
files to only have the certificate data for the public endpoint's
certificate instead of also including a self-signed certificate.

This fix is needed so that we can use k8s to manage installation of the
CSI Cinder components and later on use it for managing other Kubernetes
objects.

With Ansible 2.7, it is also possible to specify an ssl_ca_cert option
for k8s to make it work. While this option is available in 2.6 as well,
it does not seem to work.