Don't check if SSH connection works on deprovision
We were previously checking to see if an SSH connection to the first
master node works when deprovisioning. Instead of doing this, let the
deprovision task fail if a master node cannot be contacted. If we can't
reach a master node during deprovisioning, we want this to be a failure,
as this means that we'll probably need to manually clean up the cluster
to get rid of any left over dynamic Cinder volumes.

We must be able to access the API to get information about which PVs are
mapped to which Cinder volumes for cleanup. If we can't get this
mapping, we don't want deprovisioning to silently clear out the
resources that it can clear up, as this could potentially leave some
Cinder volumes behind while also destroying the information about how
these Cinder volumes are mapped to PVs.