Fix the projects view for docker registry
These changes address the problem where users are able to see all
of the docker-registry projects with anonymous pulls enabled as
part of their personal project list.

This is done by modifying the registry-viewer cluster role in such
a way that we allow access to images while preventing from getting
the associated namespaces.

The reconciliation protection has been explicitely enabled so that
restarting the masters does not reset the changes.