Make deprovisioning more robust
Reintroduce the SSH check in deprovisioning to work around an Ansible
issue where failed_when: false doesn't work when the SSH connection does
not go through. Use the result of this check to determine whether we
should talk to the API or not and later on in the playbook also use it
in an assert that prints out a helpful message in case the SSH
connection failed.

If the SSH connection did go through but the API call was unsuccesful,
show a different message to differentiate between the case when SSH
connections don't work at all and when they do work but e.g. oc is not
installed or the API is unresponsive.

After making assertions about the success of previous tasks, always
clear the inventory cache afterwards to ensure we don't have a stale
cache on subsequent runs of the deprovisioning playbook.