Fix SC delete task when we have non-active SCs
When we had storage classes set as not active, the task that deletes
storage classes that could not be updated would fail because item.failed
was not set because the previous task to update the storage classes
would skip non-active storage classes. Fix this by setting a default of
"false" for item.failed when we delete storage classes that could not be
updated.