OSO QA: Don't create namespaces for egress IPs
Since OSO QA is a permanent environment and not ephemeral like CI or
devel environments, we should not try to create egress IP namespaces
there. It looks like trying to do so will fail in the pipeline as the
requester field is immutable. The playbook for creating these egress IP
namespaces should only be used in ephemeral environments.