Add mitogen to poc-deployer
Mitogen is a plugin for Ansible that can dramatically improve
performance. Add it to the deployment container for testing in
development environments. Also add commented out lines to ansible.cfg
that can be enabled to try out Mitogen to find out any potential issues
in using it.

Mitogen's website is here: https://networkgenomics.com/ansible/