Fix URLs in web console config
The config for the web console had URLs in the openshift_public_hostname
domain that should have instead used the app domain. These were the
admin console and Hawkular metrics URLs. Fix these URLs in a way that
they work both in environments that don't have an app domain and ones
that do.

So instead of:

{{ openshift_public_hostname }}

use

{{ openshift_app_domain|default(openshift_public_hostname) }}

Another question is whether the admin console or Hawkular metrics should
be in the app domain in the first place, but since that's where they are
now, we need to implement this config change as a quick fix.