emptyDir volumes quota
This commit enables setting up quota for emptyDir volumes in all
compute nodes. This quota is enforced through the XFS quota
subsystem per FSGroup. Note: This quota is local to each compute
node; and does not limit the usage at a cluster-wide level.

Having the XFS gquota mount option in place is a requirement.