Use Swift for the docker registry
- Update the `registry_config` playbook by removing the tasks
which are related to the PVC storage configuration. The orphan
tasks were also deleted since the relevant bug has been addressed.

- Write a data migration script and run it in a form of a job.
This script copies the images data from a local volume and into
a Swift container in order to preserve existing customer images.

- The migration procedure is handled by `migrate_images_to_swift`
playbook. It first runs and waits for the data migration job to
finish, reconfigures the registry to use object storage and
finally applies the customizations from POC. The result should be
a fresh docker-registry deployment with a new storage backend
which has the old images. The expected downtime for the registry
is about 1 minute.

- Delete the Swift container when deprovisioning environments. the
OpenStack CLI had to be used since the `os_object` module does not
support recursive deletion of objects in a container.