Registry images quotas and limits
Currently all Rahti users are not limited in terms of container
images amd image streams. This commit sets a quota of 10 image
streams per namespace.

Moreover, we set a maximum size of the images. The billing model
can be according to the number of images accross all the image
streams in a namespace (Image size can be taken into account too).