Support pipeline scaleups
This commit addresses the issue of POC_INSTALLED being set on non
installed hosts when site (the pipeline) is ran. The reason behind
this is that site.yml only calls the openshift-ansible code if it
is an installation of a new cluster (and not a scaleup of an
existing one).

The proposed fix fuses site,yml and site_scaleup.yml into one
playbook which serves both purposes with some conditionals. Those
conditionals are there in order to detect whether it is a new
installation or a scaleup of an existing one:

  - First case `size(new_masters)==size(masters)`: This means that
there is no pre-existing installed master nodes, thus making this
case an installation of a fresh cluster -> run install.yml.
  - Second case `size(new_masters)!=size(masters)`: This means
that there are one or more master nodes which are installed and
running. This would refer to scaling up an existing cluster ->
run scaleup.yml.

The documentation has been updated to reflect those changes.