Ensure root when copying vault pass to RAM disk
To correctly set the file access rights for the vault pass file for
Docker the read_vault_pass_from_clipboard.bash script must run as root.
If it does not run as root then the deployment container won't execute
correctly as it cannot read the vault pass.

Add a check for root access at the start of the script so that it is
more obvious when you accidentally run the script as a non-root user.
This is mainly for usability so that one does not waste time
troubleshooting why the deployment container won't launch when the
script has been run as a non-root user.