Infra deployment node selectors
Now that we have infra nodes in place, let's move the relevant
deployments to those in order to free capacity from the master
nodes. In this commit we are chaging the node selectors for the
following:

  * heketi-storage
  * glusterblock-storage-provisioner-dc
  * docker-registry
  * registry-console
  * grafana
  * heketi-metrics-exporter
  * kube-state-metrics
  * prometheus
  * registry-image-pruner

The OpenShift console is still running on master nodes, but of
course it can also be moved.