Support external GlusterFS
Add a new role `external_glusterfs_node` for setting up non native
OpenShift GlusterFS nodes. Thia role is ran on pre_install, it
installs GlusterD and sets up a cluster. This setup is already
deployed in Rahti.

Increase the Ansible facts check timeout to 20 seconds as some
playbooks were failing because of that.