Cleanup stale GlusterFS volumes
Add a playbook under playbooks/ops which cleans up the stale
GlusterFS volumes and LVs. It keeps all the Gluster volumes which
have a relevant PV + heketidbstorage and cleansup everything else.

This should compatible with both native/non-native glusterfs
deployments and has a noop switch which needs to be set to false
for the playbook to be effective.