New cluster monitoring setup/architecture
This commits is about our new monitoring setup for our production
systems. It is mainly comprised of:
  - Prometheus with Thanos sidecar container.
  - Thanos query.
  - Thanos store.
  - Thanos compactor.
  - node-exporter.
  - kube-state-metrics.
  - Grafana.

This is suitable for long term storage since we are now storing
data on object storage. (This elimiates the need for InfluxDB).

Code related to InfluxDB has been omited.

Added documentation about the different deployments.