Support ingress IPs
This commit adds support for ingress IPs in OKD by having a
a dedicated ingress host which has the port with the dedicated
IP/Floating IP.

A new ingress stack with a single server is defined. This is a
Jinja2 template stack since multiple ports will need to be created
according to the list of ingress IPs.

Each port is essentially a Neutron port with a fixed IP (belongs
to a dedicated CIDR) and has a Floating (public) IP assigned to it.
This CIDR is sepecified under the master-config.yaml file.

The information regarding ingress ports is defined in a list
dubbed "ingress_ips", which would be defined as follows:

 ingress_ips:
  - private_ip: "192.168.1.24"
    floating_ip: "<public_ip>"
    protocol: "tcp"
    port_range_min: "3300"
    port_range_max: "3307"
    remote_ip_prefix: " 193.166.1.0/24"

This is still WIP.