Added cinder volume check to bastion
This check checks if volume attachments are out of sync between nova and
cinder. It also checks if there are volumes in error state.

This change adds the openrc to the bastion host, so that the bastion
host can talk to openstack. In practice this doesn't affect the
security, as the m2m user account password was already on the bastion.