First draft of making CI pipeline parallel
This relies on detecting gitlab environment by reading environment
variable CI_PIPELINE_ID and then overriding default environment
name with the id suffix.

So far the initial environment setup seems to be working but
site.yml fails on provisioning as the network request overlaps
with previous provisioning. If we would like to proceed with this,
we need to expand the IP ranges for these parallel deployments.