Modify monitoring memory check to check available memory
We have had memory check in place for a while and it gives false
alarms on memory checks. Checking those alarms revealed systems
have enough available memory but check was not noticing it.
This commit changes the check to read system available memory
instead of "free+cache+buff" combo that was used before.